backend/requirements.txt
# FastAPI Backend - Physical AI RAG Chatbot
# Python 3.12+
# Install with: uv pip install -r requirements.txt

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.1.1

# Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
httpx==0.26.0

# Database - PostgreSQL (Neon)
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector Store - Qdrant
qdrant-client==1.7.3

# OpenAI
openai==1.10.0

# CORS and Security
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.1

# Local Database & Auth
aiosqlite==0.21.0
passlib[bcrypt]==1.7.4
